requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
jinja2==3.1.3
python-multipart==0.0.9
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.1psutil==5.9.8
requests==2.31.0
